+++ src/s3/actions/listobjectversions.rs
@@ -1,6 +1,6 @@
 use crate::{
     s3::actions::{response_error, Action},
-    s3::responses::ListAllMyBucketsResult,
+    s3::responses::ListVersionsResult,
     s3::{request, tools, S3},
 };
 use anyhow::{anyhow, Result};
@@ -8,36 +8,38 @@ use reqwest::Method;
 use serde_xml_rs::from_str;
 use std::collections::BTreeMap;
 
-#[derive(Debug, Default)]
-pub struct ListBuckets {
-    pub max_buckets: Option<String>,
+#[derive(Debug)]
+pub struct ListObjectVersions<'a> {
+    prefix: &'a str,
 }
 
-impl ListBuckets {
+impl<'a> ListObjectVersions<'a> {
     #[must_use]
-    pub const fn new(max_buckets: Option<String>) -> Self {
-        Self { max_buckets }
+    pub const fn new(prefix: &'a str) -> Self {
+        Self { prefix }
     }
 
     /// # Errors
     ///
     /// Will return `Err` if can not make the request
-    pub async fn request(&self, s3: &S3) -> Result<ListAllMyBucketsResult> {
+    pub async fn request(&self, s3: &S3) -> Result<ListVersionsResult> {
         let (url, headers) = &self.sign(s3, tools::sha256_digest("").as_ref(), None, None)?;
 
         let response =
             request::request(url.clone(), self.http_method()?, headers, None, None, None).await?;
+
         if response.status().is_success() {
-            let buckets: ListAllMyBucketsResult = from_str(&response.text().await?)?;
-            Ok(buckets)
+            let objects: ListVersionsResult = from_str(&response.text().await?)?;
+
+            Ok(objects)
         } else {
             Err(anyhow!(response_error(response).await?))
         }
     }
 }
 
-// https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListBuckets.html
-impl Action for ListBuckets {
+// https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListObjectsV2.html
+impl Action for ListObjectVersions<'_> {
     fn http_method(&self) -> Result<Method> {
         Ok(Method::from_bytes(b"GET")?)
     }
@@ -52,15 +54,16 @@ impl Action for ListBuckets {
 
     // URL query pairs
     fn query_pairs(&self) -> Option<BTreeMap<&str, &str>> {
-        self.max_buckets.as_ref().map(|max_buckets| {
-            // URL query_pairs
-            let mut map: BTreeMap<&str, &str> = BTreeMap::new();
+        // URL query_pairs
+        let mut map: BTreeMap<&str, &str> = BTreeMap::new();
 
-            // max-buckets
-            map.insert("max-buckets", max_buckets);
+        // list-type parameter that indicates version 2 of the API
+        map.insert("versions", "");
 
-            map
-        })
+        // prefix parameter
+        map.insert("prefix", self.prefix);
+
+        Some(map)
     }
 }
 
@@ -74,7 +77,7 @@ mod tests {
 
     #[test]
     fn test_method() {
-        let action = ListBuckets::new(None);
+        let action = ListObjectVersions::new("key");
         assert_eq!(Method::GET, action.http_method().unwrap());
     }
 
@@ -90,40 +93,16 @@ mod tests {
             false,
         );
 
-        let action = ListBuckets::new(None);
+        let action = ListObjectVersions::new("key");
         let (url, headers) = action
             .sign(&s3, tools::sha256_digest("").as_ref(), None, None)
             .unwrap();
+
         assert_eq!(
-            "https://s3.us-west-1.amazonaws.com/awsexamplebucket1",
+            "https://s3.us-west-1.amazonaws.com/awsexamplebucket1?prefix=key&versions=",
             url.as_str()
         );
-        assert!(headers
-            .get("authorization")
-            .unwrap()
-            .starts_with("AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE"));
-    }
 
-    #[test]
-    fn test_max_buckets() {
-        let s3 = S3::new(
-            &Credentials::new(
-                "AKIAIOSFODNN7EXAMPLE",
-                &SecretString::new("wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY".into()),
-            ),
-            &"us-west-1".parse::<Region>().unwrap(),
-            None,
-            false,
-        );
-
-        let action = ListBuckets::new(Some("10".to_string()));
-        let (url, headers) = action
-            .sign(&s3, tools::sha256_digest("").as_ref(), None, None)
-            .unwrap();
-        assert_eq!(
-            "https://s3.us-west-1.amazonaws.com/?max-buckets=10",
-            url.as_str()
-        );
         assert!(headers
             .get("authorization")
             .unwrap()