+++ src/s3/actions/mod.rs
@@ -16,6 +16,10 @@ pub use self::listbuckets::ListBuckets;
 mod listobjectsv2;
 pub use self::listobjectsv2::ListObjectsV2;
 
+// <https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListObjectVersions.html>
+mod listobjectversions;
+pub use self::listobjectversions::ListObjectVersions;
+
 // <https://docs.aws.amazon.com/AmazonS3/latest/API/API_HeadObject.html>
 mod headobject;
 pub use self::headobject::HeadObject;